+++ examples/minimal/beta/application/beta_service.py
@@ -25,4 +25,4 @@ class BetaService(ts.ApplicationService):
         except errors.DomainError as domain_error:
             raise client.Rejected(domain_error.code, domain_error.message) from domain_error
         has_key_response = self._key_repository.has_key(MapToHasKeyRequest(key))
-        return client.CheckKeyResponse(held=has_key_response.held.value)
+        return client.CheckKeyResponse(held=has_key_response.outcome.value)