+++ layout/repo/client/__init__.py
@@ -1,5 +1,5 @@
-from repo.client.client import CheckRequest as CheckRequest
-from repo.client.client import CheckResponse as CheckResponse
+from repo.client.client import CheckLayoutRequest as CheckLayoutRequest
+from repo.client.client import CheckLayoutResponse as CheckLayoutResponse
 from repo.client.client import RepoClient as RepoClient
-from repo.client.client import TreesRequest as TreesRequest
-from repo.client.client import TreesResponse as TreesResponse
+from repo.client.client import ListTreesRequest as ListTreesRequest
+from repo.client.client import ListTreesResponse as ListTreesResponse